Move GlobalPool to custom
